improve e2e emulated tests readability
