+++ test/e2e-emulated/infra_setup.go
@@ -1,35 +1,50 @@
-/*
-Copyright 2025.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package e2esaturation
+package e2eemulated
 
 import (
 	"context"
 	"fmt"
 	"os"
 	"os/exec"
-	"testing"
 	"time"
 
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
+	. "github.com/onsi/ginkgo/v2" // nolint:all
+	. "github.com/onsi/gomega"    // nolint:all
+	promoperator "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
+	"k8s.io/client-go/kubernetes"
+	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	"github.com/llm-d-incubation/workload-variant-autoscaler/api/v1alpha1"
 	"github.com/llm-d-incubation/workload-variant-autoscaler/test/utils"
+)
+
+func init() {
+	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
+	utilruntime.Must(v1alpha1.AddToScheme(scheme))
+	utilruntime.Must(promoperator.AddToScheme(scheme))
+}
+
+// Kind cluster configuration constants
+const (
+	maximumAvailableGPUs = 4
+	numNodes             = 3
+	gpuTypes             = "mix"
+)
+
+// Kubernetes resource constants
+const (
+	controllerNamespace           = "workload-variant-autoscaler-system"
+	controllerMonitoringNamespace = "workload-variant-autoscaler-monitoring"
+	llmDNamespace                 = "llm-d-sim"
+	gatewayName                   = "infra-sim-inference-gateway-istio"
+	WVAConfigMapName              = "workload-variant-autoscaler-variantautoscaling-config"
+	saturationConfigMapName       = "workload-variant-autoscaler-saturation-scaling-config"
 )
 
 var (
@@ -46,26 +61,20 @@ var (
 	projectImage = "ghcr.io/llm-d-incubation/workload-variant-autoscaler:0.0.1-test"
 
 	MinimumReplicas = 1
+
+	k8sClient *kubernetes.Clientset
+	crClient  client.Client
+	scheme    = runtime.NewScheme()
 )
 
-const (
-	maximumAvailableGPUs = 4
-	numNodes             = 3
-	gpuTypes             = "mix"
-)
+// setupInfrastructure handles the infrastructure setup for e2e tests.
+// This includes building the image, deploying to Kind, and configuring the controller.
+func setupInfrastructure() {
+	var err error
 
-// TestSaturationE2E runs the end-to-end (e2e) test suite for saturation-based (saturation-based) mode.
-// This suite validates reactive saturation-based scaling behavior using KV cache and queue metrics.
-func TestSaturationE2E(t *testing.T) {
-	RegisterFailHandler(Fail)
-	_, _ = fmt.Fprintf(GinkgoWriter, "Starting workload-variant-autoscaler saturation-based test suite\n")
-	RunSpecs(t, "e2e saturation suite")
-}
-
-var _ = BeforeSuite(func() {
 	By("building the manager(Operator) image")
 	cmd := exec.Command("make", "docker-build", fmt.Sprintf("IMG=%s", projectImage))
-	_, err := utils.Run(cmd)
+	_, err = utils.Run(cmd)
 	ExpectWithOffset(1, err).NotTo(HaveOccurred(), "Failed to build the manager(Operator) image")
 
 	By("exporting environment variables for deployment")
@@ -79,6 +88,27 @@ var _ = BeforeSuite(func() {
 
 	initializeK8sClient()
 
+	waitForControllerReady()
+	setupCertManager()
+}
+
+// teardownInfrastructure handles the infrastructure teardown for e2e tests.
+// This includes uninstalling CertManager and destroying the Kind cluster.
+func teardownInfrastructure() {
+	// Teardown CertManager after the suite if not skipped and if it was not already installed
+	if !skipCertManagerInstall && !isCertManagerAlreadyInstalled {
+		_, _ = fmt.Fprintf(GinkgoWriter, "Uninstalling CertManager...\n")
+		utils.UninstallCertManager()
+	}
+
+	// Destroy the Kind cluster
+	cmd := exec.Command("bash", "deploy/kind-emulator/teardown.sh")
+	_, err := utils.Run(cmd)
+	ExpectWithOffset(1, err).NotTo(HaveOccurred(), "Failed to destroy Kind cluster")
+}
+
+// waitForControllerReady waits for the controller-manager pods to be ready and acquire lease.
+func waitForControllerReady() {
 	// Waiting for the workload-variant-autoscaler pods to be ready and for leader election
 	By("waiting for the controller-manager pods to be ready")
 	Eventually(func(g Gomega) {
@@ -104,7 +134,10 @@ var _ = BeforeSuite(func() {
 			g.Expect(*lease.Spec.HolderIdentity).To(ContainSubstring("controller-manager"), "Lease holderIdentity is not correct")
 		}
 	}, 2*time.Minute, 1*time.Second).Should(Succeed())
+}
 
+// configureController configures the controller by updating ConfigMaps and restarting pods.
+func configureController(KvCacheThreshold, QueueLengthThreshold, kvSpareTrigger, queueSpareTrigger float64) {
 	// Verify configuration for saturation-based mode
 	By("verifying ConfigMap is accessible")
 	cm, err := k8sClient.CoreV1().ConfigMaps(controllerNamespace).Get(context.Background(), WVAConfigMapName, metav1.GetOptions{})
@@ -132,7 +165,11 @@ queueSpareTrigger: %.2f`, KvCacheThreshold, QueueLengthThreshold, kvSpareTrigger
 
 	_, _ = fmt.Fprintf(GinkgoWriter, "Updated saturation-scaling-config with relaxed thresholds: kvCache=%.2f, queue=%.2f, kvSpare=%.2f, queueSpare=%.2f\n", KvCacheThreshold, QueueLengthThreshold, kvSpareTrigger, queueSpareTrigger)
 
-	// Restart controller pods to pick up new saturation-scaling configuration
+	restartControllerPods()
+}
+
+// restartControllerPods restarts the controller-manager pods to pick up new configuration.
+func restartControllerPods() {
 	By("restarting controller-manager pods to load new saturation configuration")
 	podList, err := k8sClient.CoreV1().Pods(controllerNamespace).List(context.Background(), metav1.ListOptions{
 		LabelSelector: "app.kubernetes.io/name=workload-variant-autoscaler",
@@ -157,7 +194,10 @@ queueSpareTrigger: %.2f`, KvCacheThreshold, QueueLengthThreshold, kvSpareTrigger
 	}, 2*time.Minute, 1*time.Second).Should(Succeed())
 
 	_, _ = fmt.Fprintf(GinkgoWriter, "Controller pods restarted and running with new saturation configuration\n")
+}
 
+// setupCertManager sets up CertManager if not already installed.
+func setupCertManager() {
 	// The tests-e2e are intended to run on a temporary cluster that is created and destroyed for testing.
 	// To prevent errors when tests run in environments with CertManager already installed,
 	// we check for its presence before execution.
@@ -172,17 +212,33 @@ queueSpareTrigger: %.2f`, KvCacheThreshold, QueueLengthThreshold, kvSpareTrigger
 			_, _ = fmt.Fprintf(GinkgoWriter, "WARNING: CertManager is already installed. Skipping installation...\n")
 		}
 	}
-})
+}
 
-var _ = AfterSuite(func() {
-	// Teardown CertManager after the suite if not skipped and if it was not already installed
-	if !skipCertManagerInstall && !isCertManagerAlreadyInstalled {
-		_, _ = fmt.Fprintf(GinkgoWriter, "Uninstalling CertManager...\n")
-		utils.UninstallCertManager()
+// initializeK8sClient initializes the Kubernetes client for testing
+func initializeK8sClient() {
+	cfg, err := func() (*rest.Config, error) {
+		if kubeconfig := os.Getenv("KUBECONFIG"); kubeconfig != "" {
+			return clientcmd.BuildConfigFromFlags("", kubeconfig)
+		}
+		return rest.InClusterConfig()
+	}()
+	if err != nil {
+		Skip("failed to load kubeconfig: " + err.Error())
 	}
 
-	// Destroy the Kind cluster
-	cmd := exec.Command("bash", "deploy/kind-emulator/teardown.sh")
-	_, err := utils.Run(cmd)
-	ExpectWithOffset(1, err).NotTo(HaveOccurred(), "Failed to destroy Kind cluster")
-})
+	// Suppress warnings to avoid spam in test output
+	cfg.WarningHandler = rest.NoWarnings{}
+
+	k8sClient, err = kubernetes.NewForConfig(cfg)
+	if err != nil {
+		Skip("failed to create kubernetes client: " + err.Error())
+	}
+
+	// Initialize controller-runtime client for custom resources
+	crClient, err = client.New(cfg, client.Options{
+		Scheme: scheme,
+	})
+	if err != nil {
+		Skip("failed to create controller-runtime client: " + err.Error())
+	}
+}
