+++ test/utils/e2eutils.go
@@ -401,7 +401,7 @@ func GetProjectDir() (string, error) {
 	}
 
 	// Handle both test packages
-	m := regexp.MustCompile(`/test/(e2e-saturation-based|e2e)`)
+	m := regexp.MustCompile(`/test/(e2e-emulated|e2e)`)
 	wd = m.ReplaceAllString(wd, "")
 	return wd, nil
 }
